Generalize hooks (#127)
* preliminary cleanup

* linting

* Update submodule reference back to dev

---------
